chore: add .npmrc file (#804)
Due to git.corp and other registries required to work with across the
corpo, it would be useful to have a `.npmrc` file specified in the
project to override any local customizations 🤔

Alternative solution, add `.npmrc` to `.gitignore` so that devs can
create it themselves when needed.




Please ensure your pull request adheres to the following guidelines:
- [x] make sure to link the related issues in this description
- [x] when merging / squashing, make sure the fixed issue references are
visible in the commits, for easy compilation of release notes

## Related Issues


Thanks for contributing!