feat: add preview page authentication to canonical audit (#859)
Please ensure your pull request adheres to the following guidelines:
- [x] make sure to link the related issues in this description
- [x] when merging / squashing, make sure the fixed issue references are
visible in the commits, for easy compilation of release notes

## Related Issues
https://jira.corp.adobe.com/browse/SITES-31664

This PR adds support for authenticating preview pages in the canonical
audit functionality. When auditing pages on `.page` domains, the system
will now retrieve and use the appropriate authentication token.