fix: avoid creating oppty when audit fails to find sitemap path (#872)
- https://jira.corp.adobe.com/browse/SITES-31562

When a sitemap path cannot be retrieved—either from the `robots.txt`
file or by appending common paths like `sitemap.xml` or
`sitemap_index.xml`, the sitemap audit fails due to the inability to
fetch the data. This failure may be intentional, as some customers
choose not to expose their sitemaps publicly. Instead of displaying this
as an error in the UI, we should handle audit failures gracefully