feat: Filter out decorative (almost all) images from audit suggestions (#923)
## Context

We have historically filtered out ALL decorative images because EDS
authoring process includes empty alts `alt=""` to all images that don't
have a textual alt. This effectively marks them as decorative, but they
need fixing, so we allowed all decorative images to be shown as
suggestions to alleviate this.

A more correct approach while the issue is not resolved in EDS, is to
only show the decorative images that are so because they have an alt
that is empty. This will not show other decorative images that:

- Use `aria-hidden="true"`
- Use `role="presentation"`

## Related Issues

- [ASSETS-52117 | Filter out decorative (almost all) images from audit
suggestions](https://jira.corp.adobe.com/browse/ASSETS-52117)
- [ASSETS-47238 | [EPIC] Alt text audit for Project Success
Studio](https://jira.corp.adobe.com/browse/ASSETS-47238)

---------

Co-authored-by: Josep Lopez <joselopez@adobe.com>