fix: detect language from url (#925)
Add a function `detectLanguageFromUrl` to check the URL path for valid
country codes and determine the language to use.

This function looks for any valid country code in the URL path segments
(the parts between slashes, excluding empty segments and the last
segment). If a valid country code is found, it returns the corresponding
language code; otherwise, it returns null. The valid country code is
defined in `country-code.js` file.

The prompt has been updated with new rules to ensure that the AI model
generates alt-text in the specified language provided with each image.

Please ensure your pull request adheres to the following guidelines:
- [ ] make sure to link the related issues in this description
- [ ] when merging / squashing, make sure the fixed issue references are
visible in the commits, for easy compilation of release notes
- [ ] If data sources for any opportunity has been updated/added, please
update the
[wiki](https://wiki.corp.adobe.com/display/AEMSites/Data+Sources+for+Opportunities)
for same opportunity.

## Related Issues
- [ASSETS-52137 | Detect language from
url](https://jira.corp.adobe.com/browse/ASSETS-52137)

Thanks for contributing!