fix: correct typo in site average CTR variable name (#955)
Updated the variable name from 'siteAgerageCtr' to 'siteAverageCtr' in
both the handler and test files to ensure consistency and accuracy in
data handling.

Please ensure your pull request adheres to the following guidelines:
- [ ] make sure to link the related issues in this description
- [ ] when merging / squashing, make sure the fixed issue references are
visible in the commits, for easy compilation of release notes
- [ ] If data sources for any opportunity has been updated/added, please
update the
[wiki](https://wiki.corp.adobe.com/display/AEMSites/Data+Sources+for+Opportunities)
for same opportunity.

## Related Issues


Thanks for contributing!