feat: Structured data audit to validate issues from scrape data (#885)
This is a major refactor of the structured data audit to support the
following features:

* Instead of just relying on Google Search Console (GSC) to retrieve
structured data issues, the audit now analyses the structured data
extracted by WAE as part of the scraping process.
* Customers not onboarded to GSC will be able to benefits from the
audit.
* See https://github.com/adobe/spacecat-content-scraper/pull/333 for
more details.
* For this the audit relies on the new
https://github.com/adobe/structured-data-validator library.
* If a customer is onboarded to GSC, issues raised by GSC will be
additionally considered, if not already covered by the validator.
Customers that are not onboarded to GSC will not bloc
* Added prefix to all logs to facilitate monitoring of issues and
missing features. This is in particular important for the further
development of the validator.
* The audit now supports structured data in Microdata and RDFa formats,
in addition to JSON-LD.
* Refactored of audit code and unit tests.

## Related Issues
* SITES-28711 Remove dependency to Google Search Console
* SITES-28714 Support additional structured data formats

Needs to be merged together with
https://github.com/adobe/spacecat-content-scraper/pull/333.

Thanks for contributing!