feat [alt-text]: SQS Integration with SpaceCat (#1025)
We are migrating the alt-text audit worker to obtain suggestions from
Mystique via SQS.

For now, we are retaining the existing Firefall implementation in the
codebase as a fallback, so we can quickly switch back if any issues
arise during the transition. Once the Mystique integration is fully
tested and stable, we will remove the Firefall code.

Please ensure your pull request adheres to the following guidelines:
- [ ] make sure to link the related issues in this description
- [ ] when merging / squashing, make sure the fixed issue references are
visible in the commits, for easy compilation of release notes
- [ ] If data sources for any opportunity has been updated/added, please
update the
[wiki](https://wiki.corp.adobe.com/display/AEMSites/Data+Sources+for+Opportunities)
for same opportunity.

## Related Issues


Thanks for contributing!