fix: A11y modify the placement of sentCount metric computation (#1174)
The purpose of this PR is to move the counting logic of the total sent
suggestions before the logic responsible for updating the received
suggestions with enhanced issues containing remediation details. Before,
when were the suggestions counted, some of them were excluded, because
they were already marked as having guidance. Now, we count the
suggestions before any of them is marked as having guidance.

Please ensure your pull request adheres to the following guidelines:
- [ ] make sure to link the related issues in this description
- [ ] when merging / squashing, make sure the fixed issue references are
visible in the commits, for easy compilation of release notes
- [ ] If data sources for any opportunity has been updated/added, please
update the
[wiki](https://wiki.corp.adobe.com/display/AEMSites/Data+Sources+for+Opportunities)
for same opportunity.

## Related Issues


Thanks for contributing!

---------

Co-authored-by: manani <manani@adobe.com>