fix: a11y logging for missing accessibility final result files (#1255)
Currently, when no accessibility final result files are found in S3, the
system always logs this as an error with [A11yProcessingError] label.
However, this is expected behavior for the first run of an accessibility
audit, causing unnecessary noise in monitoring and alerting systems.

## Related Issues
Jira: https://jira.corp.adobe.com/browse/SITES-35131

Thanks for contributing!

---------

Co-authored-by: alinalex <alexalinrauta@gmail.com>