feat: add summarization audit (#1316)
This PR adds a new audit for content summarization opportunity.
Handler: getting the top 50 pages and send to Mystique for agentic
workflow (detection + guidance)
Guidance handler: parse the message from Mystique to save the
opportunity as `generic-opportunity`

Please ensure your pull request adheres to the following guidelines:
- [ ] make sure to link the related issues in this description
- [ ] when merging / squashing, make sure the fixed issue references are
visible in the commits, for easy compilation of release notes
- [ ] If data sources for any opportunity has been updated/added, please
update the
[wiki](https://wiki.corp.adobe.com/display/AEMSites/Data+Sources+for+Opportunities)
for same opportunity.

## Related Issues


Thanks for contributing!

---------

Co-authored-by: Mark J. Becker <herzog31@users.noreply.github.com>