feat: Added additional logs for metatags (#1512)
## Description

This additional logging is needed to help with debugging and
troubleshooting while working on new features and bug fixes.

In the future, we plan to consider a more conceptual and structured
approach to logging across the entire codebase. This current
implementation serves as a temporary solution to address immediate
debugging needs.

Please ensure your pull request adheres to the following guidelines:
- [ ] make sure to link the related issues in this description
- [ ] when merging / squashing, make sure the fixed issue references are
visible in the commits, for easy compilation of release notes
- [ ] If data sources for any opportunity has been updated/added, please
update the
[wiki](https://wiki.corp.adobe.com/display/AEMSites/Data+Sources+for+Opportunities)
for same opportunity.

## Related Issues


Thanks for contributing!