LLMO-1023 (#1519) (#1520)
Please ensure your pull request adheres to the following guidelines:
- [ ] make sure to link the related issues in this description
- [ ] when merging / squashing, make sure the fixed issue references are
visible in the commits, for easy compilation of release notes
- [ ] If data sources for any opportunity has been updated/added, please
update the

[wiki](https://wiki.corp.adobe.com/display/AEMSites/Data+Sources+for+Opportunities)
for same opportunity.

## Related Issues


Thanks for contributing!

Please ensure your pull request adheres to the following guidelines:
- [ ] make sure to link the related issues in this description
- [ ] when merging / squashing, make sure the fixed issue references are
visible in the commits, for easy compilation of release notes
- [ ] If data sources for any opportunity has been updated/added, please
update the
[wiki](https://wiki.corp.adobe.com/display/AEMSites/Data+Sources+for+Opportunities)
for same opportunity.

## Related Issues


Thanks for contributing!