fix: prerender audit use scrapejobId for s3 paths (#1786)
Recently prerender content scraper was moved to `SCRAPE_CLIENT` in PR:
https://github.com/adobe/spacecat-audit-worker/pull/1772

With `SCRAPE_CLIENT`, now we are creating random `jobId` unlike
`CONTENT_SCRAPER` where we were assigning siteId as the jobId.
As a result, all files in content scraper code was pushed to a folder
with a random `jobId` while audit worker was downloading files with path
using `siteId`. Due to this mismatch, audit was not able to run the
analysis and no suggestions were created.

Updated audit worker to read content using `scrapeJobId`

Please ensure your pull request adheres to the following guidelines:
- [ ] make sure to link the related issues in this description
- [ ] when merging / squashing, make sure the fixed issue references are
visible in the commits, for easy compilation of release notes
- [ ] If data sources for any opportunity has been updated/added, please
update the
[wiki](https://wiki.corp.adobe.com/display/AEMSites/Data+Sources+for+Opportunities)
for same opportunity.

## Related Issues


Thanks for contributing!